Move BookingDirectory.java into new directory repository.
