+++ src/main/java/com/coworking_service/Controller.java
@@ -2,7 +2,7 @@ package com.coworking_service;
 
 import com.coworking_service.exception.NoSuchUserExistsException;
 import com.coworking_service.in.UserInputHandler;
-import com.coworking_service.model.BookingDirectory;
+import com.coworking_service.repository.BookingDirectory;
 import com.coworking_service.model.CoworkingSpace;
 import com.coworking_service.model.User;
 import com.coworking_service.model.UserDirectory;
